Added playerState to SaveData. Now hp and damage are saved in a file . Added a loop after dying wipes the save , so the player can start anew.

+++ Assets/Scripts/card system/model/StatMods.cs
@@ -5,10 +5,10 @@ namespace model.entity_state
     public class StatMods
     {
         public StatModType type;
-        public float amount;
+        public int amount;
         public int remainingTurns;
 
-        public StatMods(StatModType type, float amount, int remainingTurns)
+        public StatMods(StatModType type, int amount, int remainingTurns)
         {
             this.type = type;
             this.amount = amount;
