+++ Assets/Scripts/entity system/data/EntityData.cs
@@ -5,8 +5,8 @@ namespace data
     [CreateAssetMenu(fileName = "New Entity Data", menuName = "Entity/Entity Data")]
     public class EntityData : ScriptableObject
     {
-        public float health;
-        public float baseDamage;
-        public float armor;
+        public int health;
+        public int baseDamage;
+        public int armor;
     }
 }
